+++ packages/commitlint-config/index.js
@@ -1,12 +1,9 @@
 module.exports = {
-	extends: [
-		'@commitlint/config-conventional',
-	],
+	extends: ['@commitlint/config-conventional'],
 	prompt: {
 		questions: {
-			breakingBody: {
-			},
-		}
+			breakingBody: {},
+		},
 	},
 
 	rules: {