+++ packages/commitlint-config/CHANGELOG.md
@@ -5,7 +5,6 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # 1.1.0 (2023-03-09)
 
-
 ### Features
 
-* **commitlint-config:** add commitlint-config package ([40632a3](https://github.com/marcdonald/js-config/commit/40632a3ca24a919251fcb6e96990cb72659f581d))
+- **commitlint-config:** add commitlint-config package ([40632a3](https://github.com/marcdonald/js-config/commit/40632a3ca24a919251fcb6e96990cb72659f581d))